test(reference): consolidate Resize axes coverage
